Support/bin/requirements.txt
nextcloud_api_wrapper >= 0.2.1.5